% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_survey.R
\name{flatten_survey}
\alias{flatten_survey}
\title{Flatten a Qualtrics Survey}
\usage{
flatten_survey(
  survey_id,
  file_prefix = "",
  file_format = c("csv", "tsv"),
  drop_trash = TRUE
)
}
\arguments{
\item{survey_id}{string of the survey id, begins with 'SV_'}

\item{file_prefix}{string prefix to file names to be written}

\item{file_format}{one of \code{c('csv', 'tsv')}}

\item{drop_trash}{Should questions in the trash bin be discarded?}
}
\value{
a named list of the three tables generated
}
\description{
Decompose the blocks, questions, and choices of a
Qualtrics survey into rectangular data sets.
}
\author{
Sven Halvorson (svenpubmail@gmail.com)
}
