% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column_map.R
\name{get_column_map}
\alias{get_column_map}
\title{Get Column Map}
\usage{
get_column_map(survey_id, file_format = c("spss", "csv", "tsv"))
}
\arguments{
\item{survey_id}{string of the survey id, begins with 'SV_'}

\item{file_format}{one of \code{c('spss', 'csv', 'tsv')}}
}
\value{
a \code{tibble} containing information about the columns exported
by the survey associated with \code{survey_id}
}
\description{
Get Column Map
}
\details{
This function retuns a dataset describing the columns exported
by qualtrics. Some columns describe how the data export looks (ex: \code{column_exported}), others give information
about the columns (ex: \code{column_number}), and others yet give suggested
renaming conventions (ex: \code{column_harmonized}). The columns return have these meanings:
\itemize{
 \item \code{column_exported}: column names of data exported generated using
 \code{file_format}
 \item \code{column_harmonized}: suggested new name generated by combining
  \code{question_name} and \code{suffix}
 \item \code{variable_label_exported}: column label exported
 \item \code{variable_label}: column label derived
 \item \code{question_export_tag}, \code{column_export_tag}, \code{subq_export_tag} : associated export tags
 \item \code{question_id}: internal Qualtrics question id
 \item \code{import_id}: metadata from the 3rd row of CSV exports
 \item \code{question_name}: derived question name, made by using the position
 within the block.
 \item \code{suffix}: suffix created by question characteristics
 \item \code{loop_number}: loop number from "loop and merge" questions
 \item \code{column_number}: column within side-by-side questions
 \item \code{subq_number}: internal id of that sub-question
 \item \code{subq_order}: sub-question order within matrix or side-by-side questions.
 Note that this is different from \code{subq_number}
 \item \code{choice}: internal choice number from multiple choice questions
 \item \code{choice_recode}: choice recode value
 \item \code{text_entry}: binary representation of whether the choice
 allows text entry. Note that this comes from the \code{import_id}, not whether
 the column literally contains text entry data.
}
}
\note{
Uses \code{qualtables::get_responses} to fetch the exported header.
}
\author{
Sven Halvorson (svenpubmail@gmail.com)
}
