% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey.R
\name{get_survey}
\alias{get_survey}
\title{Get Information on a Qualtrics Survey}
\usage{
get_survey(survey_id, type = c("survey", "definitions", "metadata"))
}
\arguments{
\item{survey_id}{string of the survey id, begins with 'SV_'}

\item{type}{specify which API call to perform}
}
\value{
a list containing the results of the API call
}
\description{
Retrieve a JSON object that has information on a given survey.
There are currently three different api calls that can be specified by
the \code{type} argument:
\itemize{
  \item \code{survey} gives the structure of the survey in a similar
  \emph{but not identical} format as downloading the metadata
  \item \code{definitions} contains information about the survey's body such as the flow, blocks, questions ect.
  \item \code{metadata} holds information about survey settings such as the name and status
}
}
\seealso{
API call references:
\itemize{
  \item \href{https://api.qualtrics.com/73d7e07ec68b2-get-survey}{survey}
  \item \href{https://api.qualtrics.com/9d0928392673d-get-survey}{definitions}
  \item \href{https://api.qualtrics.com/a92bfbc7e58fe-get-survey-metadata}{metadata}
}
}
\author{
Sven Halvorson (svenpubmail@gmail.com)
}
