% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_info.R
\name{get_survey_info}
\alias{get_survey_info}
\title{Get Qualtrics Survey Information}
\usage{
get_survey_info(
  survey_id,
  out_path = NULL,
  type = c("survey", "definitions", "metadata")
)
}
\arguments{
\item{survey_id}{string of the survey id, begins with 'SV_'}

\item{out_path}{path to .txt file where JSON version of the return value can be written.}

\item{type}{specify which API call to perform}
}
\value{
a list containing the results of the API call
}
\description{
Retrieve a JSON object that has information on a given survey.
There are currently three different api calls that can be specified by
the \code{type} argument:
\itemize{
  \item \code{survey} gives the structure of the survye in a similar
  \emph{but not identical} format as downloading the metadata
  \item \code{definitions} contains information about the survey's body such as the flow, blocks, questions ect.
  \item \code{metadata} holds information about survey settings such as the name and status
}
}
\seealso{
API call references:
\itemize{
  \item \href{https://api.qualtrics.com/api-reference/b3A6NjExMTE-get-survey}{survey}
  \item \href{can't find this rn}{definitions}
  \item \href{https://api.qualtrics.com/api-reference/b3A6NjEwNzk-get-survey-metadata}{metadata}
}
}
\author{
Sven Halvorson (svenpubmail@gmail.com)
}
