% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad2}
\alias{pad2}
\title{Pad to two digits}
\usage{
pad2(string)
}
\arguments{
\item{string}{a string}
}
\description{
Pad strings to two digits
}
\details{
convenience wrapper for \code{stringr::str_pad}
}
