% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtreats-package.R
\docType{package}
\name{qualtreats-package}
\alias{qualtreats}
\alias{qualtreats-package}
\title{qualtreats: Process, organize, and display data from Qualtrics}
\description{
We'll describe it better when we know exactly what it does
}
\author{
\strong{Maintainer}: Sven Halvorson \email{svenpubmail@gmail.com}

}
\keyword{internal}
