% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_qtypes.R
\name{simplify_qtypes}
\alias{simplify_qtypes}
\title{Simplify Question Types}
\usage{
simplify_qtypes(survey_id)
}
\arguments{
\item{survey_id}{string of the survey id, begins with 'SV_'}
}
\value{
a \code{tibble}
}
\description{
Sometimes it's nice to categorize the questions of a Qualtrics
survey. This takes the results of \code{flatten_survey} and builds a table
that lists:
\itemize{
  \item \code{question_id}
  \item \code{column_number} for side-by-side questions
  \item \code{question_style} as one of \code{c('radio', 'checkbox', 'text', 'signature', 'descriptive')}
  \item \code{question_matrix} indicating whether the question is a matrix/side-by-side
}
}
