% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_safely}
\alias{subset_safely}
\title{Subset if possible}
\usage{
subset_safely(l, ...)
}
\arguments{
\item{l}{a named list}

\item{...}{other strings used for name accessors}
}
\description{
Subset if possible
}
\keyword{internal}
