% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_base_url}
\alias{valid_base_url}
\title{Check if a qualtrics base url is set}
\usage{
valid_base_url(base_url)
}
\arguments{
\item{base_url}{the prefix of your Qualtrics url between "https://" and ".qualtrics.com"}
}
\description{
Check if a qualtrics base url is set
}
\keyword{internal}
