% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_out_dir}
\alias{valid_out_dir}
\title{Check if a directory can be written to:}
\usage{
valid_out_dir(out_dir)
}
\arguments{
\item{out_dir}{a string to a directory}
}
\description{
Check if a directory can be written to:
}
\keyword{internal}
