% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_survey_id}
\alias{valid_survey_id}
\title{Check if survey_id matches the pattern supplied by Qualtrics}
\usage{
valid_survey_id(survey_id)
}
\arguments{
\item{survey_id}{string}
}
\description{
Check if survey_id matches the pattern supplied by Qualtrics
}
\keyword{internal}
