% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_survey.R
\name{flatten_survey}
\alias{flatten_survey}
\title{Flatten a Qualtrics Survey}
\usage{
flatten_survey(
  survey_id,
  file_format = c("spss", "csv", "tsv"),
  drop_trash = TRUE
)
}
\arguments{
\item{survey_id}{string of the survey id, begins with 'SV_'}

\item{file_format}{export format for column names. One of \code{c('spss', 'csv', 'tsv')}}

\item{drop_trash}{should questions in the trash bin be discarded?}
}
\value{
a named list of the four tables generated
}
\description{
Decompose the blocks, questions, and choices of a
Qualtrics survey into rectangular data sets.
}
\author{
Sven Halvorson (svenpubmail@gmail.com)
}
