% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_survey.R
\name{flatten_survey}
\alias{flatten_survey}
\title{Flatten a Qualtrics Survey}
\usage{
flatten_survey(
  survey_id,
  out_dir = NULL,
  file_prefix = "",
  file_format = c("csv", "tsv")
)
}
\arguments{
\item{survey_id}{string of the survey id, begins with 'SV_'}

\item{out_dir}{path to directory where files can be written}

\item{file_prefix}{string prefix to file names}

\item{file_format}{one of \code{c('csv', 'tsv')}}
}
\value{
a named list of the four tables generated
}
\description{
Decompose the blocks, questions, and choices of a
Qualtrics survey into rectangular data sets.
}
\author{
Sven Halvorson (svenpubmail@gmail.com)
TODO : make a version of subq_description that has no HTML
}
