% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_responses.R
\name{get_responses}
\alias{get_responses}
\title{Get Qualtrics Responses}
\usage{
get_responses(
  survey_id,
  out_dir = NULL,
  file_format = c("spss", "csv", "tsv"),
  useLabels = FALSE
)
}
\arguments{
\item{survey_id}{string}

\item{out_dir}{a string path to a directory for output}

\item{file_format}{one of \code{c('spss', 'csv', 'tsv')}}

\item{useLabels}{Should value labels be downloaded instead of recodes?}
}
\value{
a \code{tibble} of the responses
}
\description{
Get Qualtrics Responses
}
\author{
Sven Halvorson (svenpubmail@gmail.com)
}
