% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtreats-package.R
\docType{package}
\name{qualtreats-package}
\alias{qualtreats}
\alias{qualtreats-package}
\title{qualtreats: Get data on Qualtrics surveys from the API}
\description{
Creates formatted versions of JSON data from the Qualtrics API.
}
\author{
\strong{Maintainer}: Sven Halvorson \email{svenpubmail@gmail.com}

}
\keyword{internal}
