% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_qtypes.R
\name{simplify_qtypes}
\alias{simplify_qtypes}
\title{Simplify Question Types}
\usage{
simplify_qtypes(survey_id, flattened_survey)
}
\arguments{
\item{survey_id}{string of the survey id, begins with 'SV_'}

\item{flattened_survey}{the result of \code{qualtreats::flatten_survey}}
}
\value{
a \code{tibble}
}
\description{
Sometimes it's nice to categorize the questions of a Qualtrics
survey. This takes the results of \code{flatten_survey} and builds a table
that lists:
\itemize{
  \item \code{question_id}: internal Qualtrics question id
  \item \code{column_number}: for side-by-side questions
  \item \code{question_style}: as one of \code{c('radio', 'checkbox', 'text', 'signature', 'descriptive')}
  \item \code{question_matrix}: indicating whether the question is a matrix. Note that all
  side-by-side questions are considered matrices even if they have only one column.
  \item \code{question_sbs}: indicating whether the question is a side-by-side
  \item \code{question_loop}: marks whether the question is in a block that
  has the "loop and merge" feature enabled.
}
}
\note{
Only one of \code{survey_id} or \code{flattened_survey} is necessary.
Providing a flattened survey is offered just to save computational time.
}
