% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_safely}
\alias{subset_safely}
\title{Subset if possible}
\usage{
subset_safely(l, name)
}
\arguments{
\item{l}{a named list}

\item{name}{the name within \code{l} to try}
}
\description{
Subset if possible
}
\keyword{internal}
