% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_lab_dict.R
\name{var_lab_dict}
\alias{var_lab_dict}
\title{Apply variable names}
\usage{
var_lab_dict(df, column_names, var_labs)
}
\arguments{
\item{df}{A data frame}

\item{column_names}{character vector of column names in \code{df}}

\item{var_labs}{corresponding character vector of variable labels}
}
\value{
a data frame
}
\description{
Dictionary style application of variable names
using \code{haven::labelled}
}
